fix: app layout
